rip out default_converter method from MCPStreamableHttpKnowledgeSource
